Generated TestReport3
Generated a test report for coverage of Project 3 after editing the test
classes that CodePro generated. Was able to achieve 93% coverage. Only
classes to not reach 100% coverage were the TestAll, VendingMachineTest,
VendingMachineItemTest.

Within the VendingMachineTest & VendingMachineItemTest classes the only
methods to not reach 100% coverage were those that expected exceptions
to be thrown.